pyproject.toml
[project]
name = "tweet-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex Nam",email = "nambak80@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
